+++ packages/integrations/fastify/src/plugin.ts
@@ -59,7 +59,7 @@ const fastifyLogLayerPlugin: FastifyPluginAsync<FastifyLogLayerConfig> = async (
     requestId: requestIdConfig = true,
     autoLogging: autoLoggingConfig = true,
     contextFn,
-    group: groupConfig = true,
+    group: groupConfig,
   } = config;
   const { nameGroup, requestGroup, responseGroup } = resolveGroupConfig(groupConfig);